refactoring to remove dups and move common network static methods to ApiRequest core (out of upload)
